Added webhook for Quests/Rewards

+++ sql.sql
@@ -54,4 +54,4 @@ CREATE TABLE IF NOT EXISTS `nests` (
 // Required for Quests
 //--------------------------------------
 ALTER TABLE pokestops
-ADD (quest_id TINYINT(4), reward_id TINYINT(4));
+ADD (quest_id SMALLINT(4), reward_id SMALLINT(4));
